move SourceMapDisabledError up into errors.py
