+++ pyteal/compiler/sourcemap.py
@@ -441,18 +441,6 @@ _STATUS_CODE = "Sourcemap Status Code"
 _STATUS = "Sourcemap Status"
 
 
-# TODO: move this to pyteal.errors
-class SourceMapDisabledError(RuntimeError):
-    msg = value = """
-    Cannot calculate Teal to PyTeal source map because stack frame discovery is turned off.
-
-    To enable source maps, set `enabled = True` in `pyteal.ini`'s [pyteal-source-mapper] section.
-    """
-
-    def __str__(self):
-        return self.msg
-
-
 class TealMapItem(PyTealFrame):
     def __init__(
         self,
